requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==43.0.0
Django==5.0.7
django-ninja==1.2.2
django-ninja-extra==0.21.2
django-ninja-jwt==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
injector==0.22.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
